Fix numerous escaping issues in Attributes.
The lexer no longer breaks word tokens apart when an escape is
encountered. This will allow the parser to be simplified in a lot of
place, but that's not been done yet. But it does fix some places where
the parser wasn't handling the old way that the tokenizer split up
tokens.

This is the first commit of many where the parser will change to setting
values to the unescaped value and handle escaping those values back into
css automatically. It used to be that the API consumer had to handle
escapes on their own and that's not cool.

The biggest impact to this strategy is on the Attribute selector where
escapes are not uncommon in attribute names and quoted values are a type
of escape.

The parser now sets namespace, name  to the unescaped value.
The parser does not yet set the value to an unescaped value, that change
will follow.

The `raws` value will track the escaped value. Originally, it tracks the
exact escape sequences that were in the source, but once a value
changes, the raws value is updated and may be different. In general our
escape strategy tries to build the smallest escapes for a given
character.

The Attribute selector is much easier to mutate now, and there are
helper methods to work with the quoteMark and update the value and quote
mark at the same time. Where necessary setters on attribute are used to
keep raws values in sync.

A number of properties and some inputs to the constructor are deprecated now.
Setting an attribute's value to an escaped value is also deprecated,
where ambiguous, use setValue().